Fix SpreadingCrops to properly drop their seeds and items in a manner consistent with all crops.
