+++ src/main/java/net/dries007/tfc/objects/blocks/agriculture/BlockCropSimple.java
@@ -36,17 +36,6 @@ import net.dries007.tfc.util.skills.SkillType;
 @ParametersAreNonnullByDefault
 public abstract class BlockCropSimple extends BlockCropTFC
 {
-    private static final AxisAlignedBB[] CROPS_AABB = new AxisAlignedBB[] {
-        new AxisAlignedBB(0.125D, 0.0D, 0.125D, 0.875D, 0.125D, 0.875D),
-        new AxisAlignedBB(0.125D, 0.0D, 0.125D, 0.875D, 0.25D, 0.875D),
-        new AxisAlignedBB(0.125D, 0.0D, 0.125D, 0.875D, 0.375D, 0.875D),
-        new AxisAlignedBB(0.125D, 0.0D, 0.125D, 0.875D, 0.5D, 0.875D),
-        new AxisAlignedBB(0.125D, 0.0D, 0.125D, 0.875D, 0.625D, 0.875D),
-        new AxisAlignedBB(0.125D, 0.0D, 0.125D, 0.875D, 0.75D, 0.875D),
-        new AxisAlignedBB(0.125D, 0.0D, 0.125D, 0.875D, 0.875D, 0.875D),
-        new AxisAlignedBB(0.125D, 0.0D, 0.125D, 0.875D, 1.0D, 0.875D)
-    };
-
     private final boolean isPickable;
 
     protected BlockCropSimple(ICrop crop, boolean isPickable)
@@ -86,55 +75,6 @@ public abstract class BlockCropSimple extends BlockCropTFC
             }
         }
         return false;
-    }
-
-    @Override
-    public void harvestBlock(World worldIn, EntityPlayer player, BlockPos pos, IBlockState state, @Nullable TileEntity te, ItemStack stack)
-    {
-        // todo: in 1.14 change to use the apply skill loot table
-        super.harvestBlock(worldIn, player, pos, state, te, stack);
-
-        ItemStack seedStack = new ItemStack(ItemSeedsTFC.get(crop));
-        ItemStack foodStack = crop.getFoodDrop(state.getValue(getStageProperty()));
-        SimpleSkill skill = CapabilityPlayerData.getSkill(player, SkillType.AGRICULTURE);
-        if (skill != null)
-        {
-            foodStack.setCount(1 + RANDOM.nextInt(2 + (int) (6 * skill.getTotalLevel())));
-            if (skill.getTier().isAtLeast(SkillTier.ADEPT) && RANDOM.nextInt(10 - 2 * skill.getTier().ordinal()) == 0)
-            {
-                seedStack.setCount(2);
-            }
-            skill.add(0.04f);
-        }
-        if (!seedStack.isEmpty())
-        {
-            InventoryHelper.spawnItemStack(worldIn, pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5, seedStack);
-        }
-        if (!foodStack.isEmpty())
-        {
-            InventoryHelper.spawnItemStack(worldIn, pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5, foodStack);
-        }
-    }
-
-    @Override
-    @Nonnull
-    protected BlockStateContainer createBlockState()
-    {
-        return new BlockStateContainer(this, getStageProperty(), WILD);
-    }
-
-    @Override
-    public void getDrops(NonNullList<ItemStack> drops, IBlockAccess world, BlockPos pos, IBlockState state, int fortune)
-    {
-        drops.clear();
-    }
-
-    @Override
-    @Nonnull
-    @SuppressWarnings("deprecation")
-    public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos)
-    {
-        return CROPS_AABB[state.getValue(getStageProperty())];
     }
 
     public static BlockCropSimple create(ICrop crop, boolean isPickable)
